refactor(entities-plugins): datakit - refactor the folder structure
