+++ packages/entities/entities-plugins/src/components/free-form/Datakit/flow-editor/node-forms/InputsRecordField.vue
