+++ packages/entities/entities-plugins/src/components/free-form/Datakit/flow-editor/node-forms/NodeFormCall.vue
@@ -23,9 +23,9 @@
 <script setup lang="ts">
 import Form from '../../../shared/Form.vue'
 import Field from '../../../shared/Field.vue'
-import HttpMethodField from '../fields/HttpMethodField.vue'
-import InputsField from '../fields/InputsField.vue'
-import { CallNodeSchema } from './mock'
+import HttpMethodField from './HttpMethodField.vue'
+import InputsField from './InputsField.vue'
+import { CallNodeSchema } from '../node/mock'
 import useI18n from '../../../../../composables/useI18n'
 import NumberField from '../../../shared/NumberField.vue'
